Jquery netart
I find out how to link several function of different objects into one
click. It was quite challenging but i figured it out. However, at the
first time, I wanted to make it looping endlessly and make the stop
button. I couldn’t find out how to looping. It is a thing that I still
have to work on